requirements.txt
numpy==1.23.5
opencv-python==4.7.0.72
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.186
streamlit==1.22.0
Pillow==9.5.0